requirements.txt
email-validator>=2.2.0
flask-jwt-extended>=4.7.1
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
reportlab>=4.4.3
werkzeug>=3.1.3
flask-migrate>=4.1.0
flask-mail>=0.10.0
flask-cors>=6.0.1
stripe>=12.3.0
sqlalchemy>=2.0.42
python-dotenv==1.0.1
numpy>=1.23.0,<2.0.0
opencv-python-headless>=4.6.0,<4.10.0
sendgrid==6.11.0
beautifulsoup4==4.12.3